requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file> -c pytorch
# platform: linux-64
numpy=1.21.2
python=3.9.7
pytorch=1.11.0=py3.9_cuda11.1_cudnn8.0.5
tensorboard=2.8.0
torchvision=0.12.0=py39_cu111
